refactor(analysis): extract synthesize stage to stages/synthesize.py
Move synthesize_enriched_analysis, _apply_strategic_recompute,
_synthesize_recommendation, _compute_scenario_probabilities, _bs_nd2,
and _compute_options_probabilities out of deep_analysis_pipeline.py.
Split across three files to stay under the 300-line hook:
  stages/synthesize.py          (233 lines)
  stages/_synthesize_helpers.py (159 lines)
  stages/_synthesize_options.py  (49 lines)
Pipeline re-exports all moved names for backward compatibility.
25 pipeline tests + 697 broader tests pass unchanged.
